Blog Content Index
Adds in the ContentIndex.astro layout
Creates ContentPreviewGrid.vue and ContentPreviewCard.vue components for displaying simple previews of Astro content collections.
Creates SiteMeta.vue component to conveniently and consistently apply multiple tags that belong to the `<head>` of a document.
Extracts several tailwind styles to base.css
Creates string.ts with a couple of utility type
functions to start with.

And a blog post in a pear tree.